+++ system/database/drivers/mysqli/mysqli_driver.php
@@ -214,6 +214,13 @@ class CI_DB_mysqli_driver extends CI_DB {
 				return ($this->db->db_debug) ? $this->db->display_error($message, '', TRUE) : FALSE;
 			}
 
+			if ( ! $this->_mysqli->set_charset($this->char_set))
+			{
+				log_message('error', "Database: Unable to set the configured connection charset ('{$this->char_set}').");
+				$this->_mysqli->close();
+				return ($this->db->db_debug) ? $this->display_error('db_unable_to_set_charset', $this->char_set) : FALSE;
+			}
+
 			return $this->_mysqli;
 		}
 
@@ -260,19 +267,6 @@ class CI_DB_mysqli_driver extends CI_DB {
 		}
 
 		return FALSE;
-	}
-
-	// --------------------------------------------------------------------
-
-	/**
-	 * Set client character set
-	 *
-	 * @param	string	$charset
-	 * @return	bool
-	 */
-	protected function _db_set_charset($charset)
-	{
-		return $this->conn_id->set_charset($charset);
 	}
 
 	// --------------------------------------------------------------------